Moved tree + rock into their respective folders
